move KubeProxyConfiguration out of componentconfig API group
