+++ pkg/apis/componentconfig/register.go
@@ -45,7 +45,6 @@ func Resource(resource string) schema.GroupResource {
 func addKnownTypes(scheme *runtime.Scheme) error {
 	// TODO this will get cleaned up with the scheme types are fixed
 	scheme.AddKnownTypes(SchemeGroupVersion,
-		&KubeProxyConfiguration{},
 		&KubeSchedulerConfiguration{},
 	)
 	return nil
