+++ pkg/proxy/apis/proxyconfig/OWNERS
@@ -1,0 +1,4 @@
+approvers:
+- thockin
+reviewers:
+- sig-network-reviewers
