+++ pkg/proxy/apis/proxyconfig/v1alpha1/register.go
@@ -22,7 +22,7 @@ import (
 )
 
 // GroupName is the group name use in this package
-const GroupName = "componentconfig"
+const GroupName = "proxyconfig"
 
 // SchemeGroupVersion is group version used to register these objects
 var SchemeGroupVersion = schema.GroupVersion{Group: GroupName, Version: "v1alpha1"}
@@ -44,7 +44,7 @@ func init() {
 
 func addKnownTypes(scheme *runtime.Scheme) error {
 	scheme.AddKnownTypes(SchemeGroupVersion,
-		&KubeSchedulerConfiguration{},
+		&KubeProxyConfiguration{},
 	)
 	return nil
 }
